Add tests for Yolov7 implementation
